Refactor GH Actions repo onboarding into one file
